backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
langchain==0.0.335
faiss-cpu==1.7.4
chromadb==0.4.15
torch==2.1.1
torchvision==0.16.1
numpy==1.26.2
Pillow==10.1.0
python-multipart==0.0.6
opencv-python==4.8.1.78
mediapipe
transformers>=4.49
huggingface_hub>=0.25.0
scikit-learn==1.3.2
pandas
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1